Issue 624 - Added a new client configuration setting, "workspaceSymbols / supportsGotoClass" (not married to that terminology or structure at all), that is used to determine whether or not the language server's implementation of "workspace/symbol" is efficient enough to support the IDE's Go To Class action. The default value for that setting is "false". I have updated all existing templates to set it explicitly with those that support the notion of types using a value of "true" and those that don't using a value of "false".
The tricky part here was bridging the gap between the IDE's "gotoClassContributor" implementation (LSPWorkspaceGotoClassContributor) and the actual place where the "workspace/symbol" requests are actually issued so that we can filter by those that are configured to support more frequent calls for types-only requests via the IDE's Go To Class action. That's accomplished by extending "WorkspaceSymbolParams" to include an additional piece of context information, specifically the symbols types that are being requested. If there's a better way to bridge this gap, I'm open to it, though this ended up being pretty clean IMO.

I'll add some comments to the pull request but otherwise wait for feedback as I imagine there will be plenty. This is, of course, just the first working iteration.